src/requirements.txt
anyio
beautifulsoup4
blinker
bs4
certifi
charset-normalizer
click
dnspython
filelock
Flask
Flask-Cors
gdown
gunicorn
h11
httpcore
idna
itsdangerous
Jinja2
MarkupSafe
oauthlib
pymongo
PySocks
python-dotenv
requests
requests-oauthlib
schedule
six
sniffio
soupsieve
tqdm
tweepy
urllib3
Werkzeug
